+++ .atom/packages/gofmt/spec/formatter-spec.js
@@ -1,12 +1,25 @@
 'use babel'
 /* eslint-env jasmine */
 
+import fs from 'fs'
+import path from 'path'
+import temp from 'temp'
+
+let nl = '\n'
+
 describe('formatter', () => {
   let mainModule = null
+  let formatter = null
 
   beforeEach(() => {
+    temp.track()
+    atom.config.set('gofmt.formatOnSave', false)
+    atom.config.set('editor.defaultLineEnding', 'LF')
+
     waitsForPromise(() => {
       return atom.packages.activatePackage('go-config').then(() => {
+        return atom.packages.activatePackage('language-go')
+      }).then(() => {
         return atom.packages.activatePackage('gofmt')
       }).then((pack) => {
         mainModule = pack.mainModule
@@ -14,7 +27,197 @@ describe('formatter', () => {
     })
 
     waitsFor(() => {
-      return mainModule.getGoconfig() !== false
+      return mainModule.goconfig
+    })
+
+    waitsFor(() => {
+      formatter = mainModule.getFormatter()
+      return formatter
+    })
+
+    waitsFor(() => {
+      return formatter.ready()
+    })
+  })
+
+  describe('when a simple file is opened', () => {
+    let editor
+    let filePath
+    let saveSubscription
+    let actual
+
+    beforeEach(() => {
+      let directory = fs.realpathSync(temp.mkdirSync())
+      atom.project.setPaths([directory])
+      filePath = path.join(directory, 'main.go')
+      fs.writeFileSync(filePath, '')
+      waitsForPromise(() => {
+        return atom.workspace.open(filePath).then((e) => {
+          editor = e
+          saveSubscription = e.onDidSave(() => {
+            actual = e.getText()
+          })
+        })
+      })
+    })
+
+    afterEach(() => {
+      if (saveSubscription) {
+        saveSubscription.dispose()
+      }
+
+      actual = undefined
+    })
+
+    describe('when format on save is disabled and gofmt is the tool', () => {
+      beforeEach(() => {
+        atom.config.set('gofmt.formatOnSave', false)
+        formatter.resetFormatterCache()
+        formatter.updateFormatterCache()
+        atom.config.set('gofmt.formatTool', 'gofmt')
+        waitsFor(() => {
+          return formatter.ready()
+        })
+      })
+
+      it('does not format the file on save', () => {
+        let text = 'package main' + nl + nl + 'func main()  {' + nl + '}' + nl
+        let expected = text
+        let formatted = 'package main' + nl + nl + 'func main() {' + nl + '}' + nl
+
+        runs(() => {
+          let buffer = editor.getBuffer()
+          buffer.setText(text)
+          buffer.save()
+        })
+
+        waitsFor(() => { return actual })
+
+        runs(() => {
+          expect(actual).toBe(expected)
+          expect(actual).not.toBe(formatted)
+        })
+      })
+
+      it('formats the file on command', () => {
+        let text = 'package main' + nl + nl + 'func main()  {' + nl + '}' + nl
+        let unformatted = text
+        let formatted = 'package main' + nl + nl + 'func main() {' + nl + '}' + nl
+
+        runs(() => {
+          let buffer = editor.getBuffer()
+          buffer.setText(text)
+          buffer.save()
+        })
+
+        waitsFor(() => {
+          return actual
+        })
+
+        runs(() => {
+          expect(actual).toBe(unformatted)
+          expect(actual).not.toBe(formatted)
+          let target = atom.views.getView(editor)
+          atom.commands.dispatch(target, 'golang:gofmt')
+        })
+
+        runs(() => {
+          expect(editor.getText()).toBe(formatted)
+        })
+      })
+    })
+
+    describe('when format on save is enabled and gofmt is the tool', () => {
+      beforeEach(() => {
+        atom.config.set('gofmt.formatOnSave', true)
+        formatter.resetFormatterCache()
+        formatter.updateFormatterCache()
+        atom.config.set('gofmt.formatTool', 'gofmt')
+        waitsFor(() => {
+          return formatter.ready()
+        })
+      })
+
+      it('formats the file on save', () => {
+        let text = 'package main' + nl + nl + 'func main()  {' + nl + '}' + nl
+        let expected = 'package main' + nl + nl + 'func main() {' + nl + '}' + nl
+
+        runs(() => {
+          let buffer = editor.getBuffer()
+          buffer.setText(text)
+          buffer.save()
+        })
+
+        waitsFor(() => {
+          return actual
+        })
+
+        runs(() => {
+          expect(actual).toBe(expected)
+        })
+      })
+    })
+
+    describe('when format on save is enabled and goimports is the tool', () => {
+      beforeEach(() => {
+        atom.config.set('gofmt.formatOnSave', true)
+        formatter.resetFormatterCache()
+        formatter.updateFormatterCache()
+        atom.config.set('gofmt.formatTool', 'goimports')
+        waitsFor(() => {
+          return formatter.ready()
+        })
+      })
+
+      it('formats the file on save', () => {
+        let text = 'package main' + nl + nl + 'func main()  {' + nl + '}' + nl
+        let expected = 'package main' + nl + nl + 'func main() {' + nl + '}' + nl
+
+        runs(() => {
+          let buffer = editor.getBuffer()
+          buffer.setText(text)
+          buffer.save()
+        })
+
+        waitsFor(() => {
+          return actual
+        })
+
+        runs(() => {
+          expect(actual).toBe(expected)
+        })
+      })
+    })
+
+    describe('when format on save is enabled and goreturns is the tool', () => {
+      beforeEach(() => {
+        atom.config.set('gofmt.formatOnSave', true)
+        formatter.resetFormatterCache()
+        formatter.updateFormatterCache()
+        atom.config.set('gofmt.formatTool', 'goreturns')
+        waitsFor(() => {
+          return formatter.ready()
+        })
+      })
+
+      it('formats the file on save', () => {
+        let text = 'package main' + nl + nl + 'func main()  {' + nl + '}' + nl
+        let expected = 'package main' + nl + nl + 'func main() {' + nl + '}' + nl
+
+        runs(() => {
+          let buffer = editor.getBuffer()
+          buffer.setText(text)
+          buffer.save()
+        })
+
+        waitsFor(() => {
+          return actual
+        })
+
+        runs(() => {
+          expect(actual).toBe(expected)
+        })
+      })
     })
   })
 })